HACK: Backdate odometry time 100ms to match LIDAR data.
Based on RViz plot, odometry and LIDAR scan data are offset by about
100ms, which confuses the mapping algorithm. Right now it's unknown
if the error comes from the LIDAR node or the Roboclaw node, never
mind where inside that node is causing the timing issue.

In the short term, the timestamp of odometry data is getting backdated
by 100ms. This is close enough to LIDAR data for gmapping to function
but is only a hack and not an actual fix.

In the long term, need to figure out where that time difference is coming
from and fix it at the source.